refactor daemon cleanup test ownership
